syntheticGeneration/requirements.txt
numpy>=1.20.0
h5py>=3.0.0
click>=8.0.0
PyYAML>=6.0.0
